requirements.txt
nextcord
python-dotenv
PyNaCl
youtube_dl
pytz
openai
Wavelink==1.3.4